feat: completed authoriztion on front and fixed backend

+++ isa-project/src/main/java/isaproject/repository/FishingTrainerRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import isaproject.model.FishingTrainer;
 
-public interface FishingTrainerRepository extends JpaRepository<FishingTrainer, Integer> {
+public interface FishingTrainerRepository extends JpaRepository<FishingTrainer, Long> {
 
 }
